feat(server): add authenticateUser middleware
Implements authenticateUser middleware and use it in authRoutes (updateUser route) and server (in
front of jobsRoute) in order to limit access to these routes only to authenticated users. Also, it
was necessary to create a special typescript file to be able expand Request object properties to
accept a new property called user.