fix(client): fix bug in AddJob page
Removes fields _id and createdAt from initialState to prevent a bug when trying to add a new job.
These fields were previously being passed to the che createJob function along with the rest of the
job information and this was not the correct behavior since these fields have to be created by the
database automatically.